tests/wpt/web-platform-tests/referrer-policy/no-referrer/attr-referrer/same-origin/http-https/iframe-tag/no-redirect/generic.http.html

<!DOCTYPE html>
<!-- DO NOT EDIT! Generated by referrer-policy/generic/tools/generate.py using referrer-policy/generic/template/test.release.html.template. -->
<html>
  <head>
    <title>Referrer-Policy: Referrer Policy is set to 'no-referrer'</title>
    <meta name="description" content="Check that sub-resource never gets the referrer URL.">
    <link rel="author" title="Kristijan Burnik" href="burnik@chromium.org">
    <link rel="help" href="https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-no-referrer">
    <meta name="assert" content="The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via iframe-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- TODO(kristijanburnik): Minify and merge both: -->
    <script src="/referrer-policy/generic/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js?pipe=sub"></script>
  </head>
  <body>
    <script>
      ReferrerPolicyTestCase(
        {
          "referrer_policy": "no-referrer",
          "delivery_method": "attr-referrer",
          "redirection": "no-redirect",
          "origin": "same-origin",
          "source_protocol": "http",
          "target_protocol": "https",
          "subresource": "iframe-tag",
          "subresource_path": "/referrer-policy/generic/subresource/document.py",
          "referrer_url": "omitted"
        },
        document.querySelector("meta[name=assert]").content,
        new SanityChecker()
      ).start();
      </script>
    <div id="log"></div>
  </body>
</html>
