LayoutTests/imported/w3c/web-platform-tests/referrer-policy/no-referrer/meta-referrer/cross-origin/http-https/iframe-tag/no-redirect/generic.http-expected.txt



PASS The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via iframe-tag using the meta-referrer
                                 delivery method with no-redirect and when
                                 the target request is cross-origin.
